+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# REPLACE_GLOBALLY_WITH_NEXT_VERSION
+# 2.1.0
 - PPI-174 - Cart and order line items are now sent with their SKU
 - PPI-174 - Added events to adjust line items which are sent to PayPal
 - PPI-202 - Fix PayPal checkout for customers with net prices